+++ internal/handlers/common/aggregations/stages/collstats.go
@@ -111,15 +111,15 @@ func (c *collStats) Process(ctx context.Context, iter types.DocumentsIterator, c
 			for _, key := range scalable {
 				path := types.NewStaticPath("storageStats", key)
 				val := must.NotFail(res.GetByPath(path))
-				must.NoError(res.SetByPath(path, val.(int32)/int32(scale)))
+				must.NoError(res.SetByPath(path, val.(int64)/int64(scale)))
 			}
 
 			path := types.NewStaticPath("storageStats", "indexSizes")
 			indexSizesDoc := must.NotFail(res.GetByPath(path)).(*types.Document)
 
 			for _, name := range indexSizesDoc.Keys() {
-				size := must.NotFail(indexSizesDoc.Get(name)).(int32)
-				indexSizesDoc.Set(name, size/int32(scale))
+				size := must.NotFail(indexSizesDoc.Get(name)).(int64)
+				indexSizesDoc.Set(name, size/int64(scale))
 			}
 
 			must.NoError(res.SetByPath(path, indexSizesDoc))
